requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
google-genai>=0.2.0
python-dotenv==1.0.1
dj-database-url==2.1.0
Pillow>=10.0.0
